backend/requirements.txt
# KidBot Backend Dependencies
# For deployment on Render.com

# =============================================================================
# Web Framework
# =============================================================================
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9
websockets==12.0
pydantic==2.6.1

# =============================================================================
# LLM & AI
# =============================================================================
openai==1.12.0

# =============================================================================
# RAG / Memory
# =============================================================================
chromadb==0.4.22
sentence-transformers==2.3.1

# =============================================================================
# Voice & Audio
# =============================================================================
# Text-to-Speech (Multi-language)
edge-tts==6.1.9

# Speech Recognition
SpeechRecognition==3.10.1

# Voice Verification (optional - comment out if not needed)
# resemblyzer==0.1.4

# =============================================================================
# Core Dependencies
# =============================================================================
numpy==1.26.4
python-dotenv==1.0.1
PyYAML==6.0.1

# =============================================================================
# Production Server
# =============================================================================
gunicorn==21.2.0
